Windows build complete...
Boost and GLUT dependencies were removed, and GUFG is building
successfully on Windows. Next step is to revise the build instructions
documentation on Google Docs. Note that there a number of DLLs you need
to include in the final zip file if you intend to distribute GUFG. This
will be detailed in the documentation eventually.