Added reversal buffer window
Basically, any move inputted that is not currently allowed by the cancel
state will check if the current move is within the REVERSAL WINDOW. This
window only exists for moves that last a total of 10 or more frames, and
reversal timing is the last 3 frames of any such move. Any move input
during this window will happen as soon as possible after the current
move finishes. This allows a very small amount of leniency (that
wouldn't otherwise exist) for things like reversals on wakeup, as well
as links in combos.

This patch also had some general cleaning up of the prepHooks() function